+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'AESP'
+rootProject.name = 'Say Better'
