--- src/main/java/com/aesp/AespApplication.java
@@ -1,13 +1,0 @@
-package com.aesp;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AespApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(AespApplication.class, args);
-    }
-
-}
