--- src/main/java/com/aesp/domain/temp/converter/TempConverter.java
@@ -1,4 +1,0 @@
-package com.aesp.domain.temp.converter;
-
-public class TempConverter {
-}
