--- src/main/java/com/aesp/domain/temp/dao/impl/TempReadRepositoryImpl.java
@@ -1,6 +1,0 @@
-package com.aesp.domain.temp.dao.impl;
-
-import com.aesp.domain.temp.dao.repository.TempRepository;
-
-public class TempReadRepositoryImpl implements TempRepository {
-}
