--- src/main/java/com/aesp/domain/temp/dao/impl/TempWriteRepositoryImpl.java
@@ -1,6 +1,0 @@
-package com.aesp.domain.temp.dao.impl;
-
-import com.aesp.domain.temp.dao.repository.TempRepository;
-
-public class TempWriteRepositoryImpl implements TempRepository {
-}
