--- src/main/java/com/aesp/domain/temp/dao/repository/TempRepository.java
@@ -1,4 +1,0 @@
-package com.aesp.domain.temp.dao.repository;
-
-public interface TempRepository {
-}
