--- src/main/java/com/aesp/domain/temp/domain/Temp.java
@@ -1,4 +1,0 @@
-package com.aesp.domain.temp.domain;
-
-public class Temp {
-}
