--- src/main/java/com/aesp/domain/temp/dto/TempRequest.java
@@ -1,4 +1,0 @@
-package com.aesp.domain.temp.dto;
-
-public class TempRequest {
-}
