--- src/main/java/com/aesp/domain/temp/dto/TempResponse.java
@@ -1,14 +1,0 @@
-package com.aesp.domain.temp.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-public class TempResponse {
-
-	@Getter
-	@Builder
-	public static class ExceptionTestDto {
-		private String message;
-	}
-
-}
