--- src/main/java/com/aesp/global/common/code/BaseCode.java
@@ -1,11 +1,0 @@
-package com.aesp.global.common.code;
-
-import com.aesp.global.common.response.ResponseDto;
-
-public interface BaseCode {
-
-	ResponseDto.ReasonDto getReason();
-
-	ResponseDto.ReasonDto getReasonHttpStatus();
-
-}
