--- src/main/java/com/aesp/global/common/code/BaseErrorCode.java
@@ -1,11 +1,0 @@
-package com.aesp.global.common.code;
-
-import com.aesp.global.common.response.ResponseDto;
-
-public interface BaseErrorCode {
-
-	ResponseDto.ErrorReasonDto getReason();
-
-	ResponseDto.ErrorReasonDto getReasonHttpStatus();
-
-}
