--- src/main/java/com/aesp/global/config/docs/SwaggerConfig.java
@@ -1,26 +1,0 @@
-package com.aesp.global.config.docs;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import io.swagger.v3.oas.models.Components;
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-
-@Configuration
-public class SwaggerConfig {
-
-	@Bean
-	public OpenAPI openAPI() {
-		return new OpenAPI()
-				.components(new Components())
-				.info(apiInfo());
-	}
-
-	private Info apiInfo() {
-		return new Info()
-				.title("Say Better API")
-				.description("API Documentation for Say Better")
-				.version("1.0.0");
-	}
-}
