+++ src/main/java/com/saybetter/SayBetterApplication.java
@@ -1,0 +1,13 @@
+package com.saybetter;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SayBetterApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(SayBetterApplication.class, args);
+    }
+
+}
