+++ src/main/java/com/saybetter/domain/temp/converter/TempConverter.java
@@ -1,0 +1,4 @@
+package com.saybetter.domain.temp.converter;
+
+public class TempConverter {
+}
