+++ src/main/java/com/saybetter/domain/temp/dao/impl/TempReadRepositoryImpl.java
@@ -1,0 +1,6 @@
+package com.saybetter.domain.temp.dao.impl;
+
+import com.saybetter.domain.temp.dao.repository.TempRepository;
+
+public class TempReadRepositoryImpl implements TempRepository {
+}
