+++ src/main/java/com/saybetter/domain/temp/dao/impl/TempWriteRepositoryImpl.java
@@ -1,0 +1,6 @@
+package com.saybetter.domain.temp.dao.impl;
+
+import com.saybetter.domain.temp.dao.repository.TempRepository;
+
+public class TempWriteRepositoryImpl implements TempRepository {
+}
