+++ src/main/java/com/saybetter/domain/temp/dao/repository/TempRepository.java
@@ -1,0 +1,4 @@
+package com.saybetter.domain.temp.dao.repository;
+
+public interface TempRepository {
+}
