+++ src/main/java/com/saybetter/domain/temp/domain/Temp.java
@@ -1,0 +1,4 @@
+package com.saybetter.domain.temp.domain;
+
+public class Temp {
+}
