+++ src/main/java/com/saybetter/domain/temp/dto/TempRequest.java
@@ -1,0 +1,4 @@
+package com.saybetter.domain.temp.dto;
+
+public class TempRequest {
+}
