+++ src/main/java/com/saybetter/domain/temp/dto/TempResponse.java
@@ -1,0 +1,14 @@
+package com.saybetter.domain.temp.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+
+public class TempResponse {
+
+	@Getter
+	@Builder
+	public static class ExceptionTestDto {
+		private String message;
+	}
+
+}
