+++ src/main/java/com/saybetter/domain/temp/exception/TempException.java
@@ -1,0 +1,10 @@
+package com.saybetter.domain.temp.exception;
+
+import com.saybetter.global.common.code.BaseErrorCode;
+import com.saybetter.global.common.exception.GeneralException;
+
+public class TempException extends GeneralException {
+	public TempException(BaseErrorCode code) {
+		super(code);
+	}
+}
