+++ src/main/java/com/saybetter/global/common/code/BaseCode.java
@@ -1,0 +1,11 @@
+package com.saybetter.global.common.code;
+
+import com.saybetter.global.common.response.ResponseDto;
+
+public interface BaseCode {
+
+	ResponseDto.ReasonDto getReason();
+
+	ResponseDto.ReasonDto getReasonHttpStatus();
+
+}
