+++ src/main/java/com/saybetter/global/common/code/BaseErrorCode.java
@@ -1,0 +1,11 @@
+package com.saybetter.global.common.code;
+
+import com.saybetter.global.common.response.ResponseDto;
+
+public interface BaseErrorCode {
+
+	ResponseDto.ErrorReasonDto getReason();
+
+	ResponseDto.ErrorReasonDto getReasonHttpStatus();
+
+}
