--- src/test/java/com/aesp/AespApplicationTests.java
@@ -1,13 +1,0 @@
-package com.aesp;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AespApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
