+++ src/test/java/com/saybetter/SayBetterApplicationTests.java
@@ -1,0 +1,13 @@
+package com.saybetter;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SayBetterApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
